rag_system/requirements.txt
anthropic>=0.25.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
google-genai
